db changes

+++ NearXServer/Migrations/20220103150856_init.cs
@@ -22,7 +22,7 @@ namespace NearXServer.Migrations
                         .Annotation("MySql:CharSet", "utf8mb4"),
                     Note = table.Column<string>(type: "longtext", nullable: false)
                         .Annotation("MySql:CharSet", "utf8mb4"),
-                    ElementId = table.Column<int>(type: "int", nullable: false)
+                    ElementId = table.Column<long>(type: "long", nullable: false)
                 },
                 constraints: table =>
                 {
