+++ NearXShared/Models/Review.cs
@@ -14,6 +14,6 @@ namespace NearXShared.Models
 
         public string Note { get; set; }
 
-        public int ElementId { get; set; }
+        public long ElementId { get; set; }
     }
 }
